v0.2.6
- finished items fabric
- testing all mainInvertoryScene

+++ Assets/Code/Items/Potions/airUpMix4U.cs
@@ -10,10 +10,10 @@ namespace Assets.Code.Items.Potions
 {
     public class AirUpMix4U : A_Item
     {
-        public short defValue;
+        public int defValue;
         public int defTime;
 
-        public AirUpMix4U(short _defVal, int _defTime, int _price) : base(_price)
+        public AirUpMix4U(int _defVal, int _defTime, int _price) : base(_price)
         {
             defValue = _defVal;
             defTime = _defTime;
