+++ Assets/Code/Items/Potions/deathUpMix1S.cs
@@ -10,10 +10,10 @@ namespace Assets.Code.Items.Potions
 {
     public class DeathUpMix1S : A_Item
     {
-        public short defValue;
+        public int defValue;
         public int defTime;
 
-        public DeathUpMix1S(short _defVal, int _defTime, int _price) : base(_price)
+        public DeathUpMix1S(int _defVal, int _defTime, int _price) : base(_price)
         {
             defValue = _defVal;
             defTime = _defTime;
