+++ Assets/Code/Items/Potions/fireUpMix3G.cs
@@ -10,10 +10,10 @@ namespace Assets.Code.Items.Potions
 {
     public class FireUpMix3G : A_Item
     {
-        public short defValue;
+        public int defValue;
         public int defTime;
 
-        public FireUpMix3G(short _defVal, int _defTime, int _price) : base(_price)
+        public FireUpMix3G(int _defVal, int _defTime, int _price) : base(_price)
         {
             defValue = _defVal;
             defTime = _defTime;
