+++ Assets/Code/Items/Potions/physUpMix1S.cs
@@ -10,10 +10,10 @@ namespace Assets.Code.Items.Potions
 {
     public class PhysUpMix1S : A_Item
     {
-        public short defValue;
+        public int defValue;
         public int defTime;
 
-        public PhysUpMix1S(short _defVal, int _defTime, int _price) : base(_price)
+        public PhysUpMix1S(int _defVal, int _defTime, int _price) : base(_price)
         {
             defValue = _defVal;
             defTime = _defTime;
