+++ Assets/Code/Items/Potions/physUpMix3G.cs
@@ -10,10 +10,10 @@ namespace Assets.Code.Items.Potions
 {
     public class PhysUpMix3G : A_Item
     {
-        public short defValue;
+        public int defValue;
         public int defTime;
 
-        public PhysUpMix3G(short _defVal, int _defTime, int _price) : base(_price)
+        public PhysUpMix3G(int _defVal, int _defTime, int _price) : base(_price)
         {
             defValue = _defVal;
             defTime = _defTime;
